move eager implementation to funsor.distribution
